修复了一堆问题，现已可用

+++ admin/delete.aspx.cs
@@ -11,6 +11,12 @@ namespace messageboard.admin
     {
         protected void Page_Load(object sender, EventArgs e)
         {
+            if (Session["admin"] == null)
+            {
+                Response.Write("<script>alert('请先登录管理员账号！');window.location='login.aspx';</script>");
+                Response.End();
+                return;
+            }
             DBClass db1 = new DBClass();
             string strsql = "delete from gbook where id=" + Request.QueryString["id"];
             db1.ExecuteSql(strsql);
